Refactor the truth table tests. Ignore some tests that are failing with NPEs - filing a new story for that.
