+++ src/test/java/org/broadinstitute/dsde/consent/ontology/truthtable/AggregateAnalysisMatchTest.java
@@ -1,10 +1,32 @@
 package org.broadinstitute.dsde.consent.ontology.truthtable;
 
+import io.dropwizard.testing.junit.ResourceTestRule;
 import org.broadinstitute.dsde.consent.ontology.datause.models.*;
 import org.broadinstitute.dsde.consent.ontology.resources.MatchPair;
+import org.broadinstitute.dsde.consent.ontology.resources.MatchResource;
+import org.glassfish.jersey.test.grizzly.GrizzlyWebTestContainerFactory;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
 import org.junit.Test;
+import org.mockito.InjectMocks;
 
 public class AggregateAnalysisMatchTest extends TruthTableTests {
+
+    @InjectMocks
+    private static MatchResource matchResource = new MatchResource();
+
+    /**
+     * Use GrizzlyTestContainerFactory to process async requests.
+     */
+    @ClassRule
+    public static final ResourceTestRule RULE = ResourceTestRule.builder()
+        .setTestContainerFactory(new GrizzlyWebTestContainerFactory())
+        .addResource(matchResource).build();
+
+    @BeforeClass
+    public static void setUp() throws Exception {
+        matchResource.setOntologyList(getOntologyListMock());
+    }
 
     private UseRestriction darAAA = new And(
         new Named("http://www.broadinstitute.org/ontologies/DURPO/aggregate_analysis"),
@@ -73,7 +95,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAA, dulUC1);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -85,7 +107,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAA, dulUC1);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -97,7 +119,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAA, dulUC2);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -109,7 +131,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAA, dulUC2);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -121,7 +143,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAA, dulUC3);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -133,7 +155,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAA, dulUC3);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -145,7 +167,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAB, dulUC1);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -157,7 +179,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAB, dulUC1);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -169,7 +191,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAB, dulUC2);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
 
@@ -183,7 +205,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAB, dulUC2);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -195,7 +217,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAB, dulUC3);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -207,7 +229,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAB, dulUC3);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -219,7 +241,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAC, dulUC1);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -231,7 +253,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAC, dulUC1);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -243,7 +265,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAC, dulUC2);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -255,7 +277,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAC, dulUC2);
-        assertResponse(getResponseFuture(pair), true);
+        assertResponse(getResponseFuture(RULE, pair), true);
     }
 
     @Test
@@ -267,7 +289,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darAAC, dulUC3);
-        assertResponse(getResponseFuture(pair), false);
+        assertResponse(getResponseFuture(RULE, pair), false);
     }
 
     @Test
@@ -279,7 +301,7 @@ public class AggregateAnalysisMatchTest extends TruthTableTests {
         // Response should be positive
 
         MatchPair pair = new MatchPair(darDefaultAAC, dulUC3);
-        assertResponse(getResponseFuture(pair), false);
+        assertResponse(getResponseFuture(RULE, pair), false);
     }
 
 }
